Crate mg_settings

Parse config files.

Usage

Call the parse function on the input.

Modules

error
Parse error type.

key
Type for representing keys and functions for parsing strings into Keys.

settings
Settings manager.

Structs

Config
The parsing configuration.

MetaData
Command/setting meta-data coming from the attributes. See EnumMetaData to see the list of supported attributes.

Parser
The config parser.

Enums

Command
The Command enum represents a command from a config file.

Value
The Value enum represents a value along with its type.

Traits

CompletionValues
Trait to specify the completion values for a type.

EnumFromStr
The EnumFromStr trait is used to specify how to construct an enum value from a string.

EnumMetaData
Tre EnumMetaData trait is used to get associated meta-data for the enum variants. The meta-data is specified using the following attributes: #[completion(hidden)] #[special_command] #[help(Command help)]

SettingCompletion
Trait specifying the value completions for settings.