Enum mg_settings::Command

pub enum Command<T> {
    App(String),
    Custom(T),
    Map {
        action: String,
        keys: Vec<Key>,
        mode: String,
    },
    Set(String, Value),
    Unmap {
        keys: Vec<Key>,
        mode: String,
    },
}

The Command enum represents a command from a config file.

Variants
App(String)

A command from the application library.

Custom(T)

A custom command.

Map

A map command creates a new key mapping.

Fields of Map

action: String

The action that will be executed when the keys are pressed.

keys: Vec<Key>

The key shortcut to trigger the action.

mode: String

The mode in which this mapping is available.

Set(String, Value)

A set command sets a value to an option.

Unmap

An unmap command removes a key mapping.

Fields of Unmap

keys: Vec<Key>

The key shortcut to remove.

mode: String

The mode in which this mapping is available.