Module mg_settings::error

Parse error type.

Structs

Error
Struct which holds information about an error at a specific position.

Enums

ErrorType
A set of error types that can occur parsing the settings file.

Type Definitions

Result
A type alias over the specific Result type used by the parser to indicate whether it is successful or not.