Module mg_settings::key

Type for representing keys and functions for parsing strings into Keys.

Enums

Key
Enum representing the keys that can be used in a mapping.

Functions

parse_keys
Parse a string into a vector of Keys.