Crate mg_settings

Parse config files.

Usage

Call the parse function on the input.

Modules

error
Parse error type.

key
Type for representing keys and functions for parsing strings into Keys.

Macros

commands

Structs

Config
The parsing configuration.

Parser
The config parser.

Enums

Command
The Command enum represents a command from a config file.

Value
The Value enum represents a value along with its type.

Traits

EnumFromStr
The EnumFromStr trait is used to specify how to construct an enum value from a string.